Pull Dashboard into own component used by page.txt
